Support removing MIDI patches again
This is a bit of an odd solution, as whilst LiSP's configuration code detects
whether a config (sub)key has been added or had its associated value changed,
it doesn't pick up on key deletions.

Thus, we null the stored device name, and just treat any device patches with a
null'd name as if it doesn't exist.

Another odd quirk with LiSP's Signal class in that it doesn't like to be
connected to a slot belonging to a Qt5 widget, due to not being able to see
the slot's method signature. Hopefully the use of PyQt5's pyqtSignal class
instead doesn't cause any problems.